Bump the rule numbers to fix conflict
